storage_box/Forked/newsnerd-telegram-bot-master/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pyaes==1.6.1
Pyrogram==2.0.100
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
TgCrypto==1.2.5
urllib3==1.26.14
